add exceptions to notifications

+++ spot-server/src/app/db/notifications.ts
@@ -54,9 +54,9 @@ function addReplyNotification(senderId: string, receiverId: string, postId: stri
     });
 }
 
-function setNotificationSeen(notificationId: string) {
-    var sql = `UPDATE notifications SET seen = true WHERE id = ?`;
-    var values = [notificationId];
+function setNotificationSeen(notificationId: string, accountId: string) {
+    var sql = `UPDATE notifications SET seen = true WHERE id = ? AND account_id = ?`;
+    var values = [notificationId, accountId];
     return db.query(sql, values);
 }
 
@@ -66,9 +66,9 @@ function setAllNotificationsSeen(accountId: string) {
     return db.query(sql, values);
 }
 
-function deleteNotificationById(id: string) {
-    var sql = `DELETE FROM notifications WHERE id = ?`;
-    var values = [id];
+function deleteNotificationById(id: string, accountId: string) {
+    var sql = `DELETE FROM notifications WHERE id = ? AND account_id = ?`;
+    var values = [id, accountId];
     return db.query(sql, values);
 }
